Move aws-operations and fs-operations out of utils/
